Fix "kind" attribute for subtitles
The official value for the "kind" attribute of the <track> element is `subtitles` instead of `subtitle`. This can also be found here: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/track
Without the correct value, the text tracks will not show in the Safari browser and do not return an error or error message. Chrome, Edge and Firefox ignore it.